fix some (all? not likely) problems with dragging close to 2^32-1 frames
git-svn-id: svn://localhost/ardour2/branches/2.0-ongoing@1924 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ gtk2_ardour/draginfo.h
@@ -40,7 +40,7 @@ struct DragInfo {
     ItemType            item_type;
     void* data;
     nframes_t last_frame_position;
-    int32_t pointer_frame_offset;
+    int64_t pointer_frame_offset;
     nframes_t grab_frame;
     nframes_t last_pointer_frame;
     nframes_t current_pointer_frame;
